+++ tests/closures.rs
@@ -239,7 +239,7 @@ fn test_closures_shared_obj() -> Result<(), Box<EvalAltResult>> {
     let f = move |p1: TestStruct, p2: TestStruct| -> Result<(), Box<EvalAltResult>> {
         let action_ptr = res["action"].clone().cast::<FnPtr>();
         let name = action_ptr.fn_name();
-        engine.call_fn::<_, ()>(&mut Scope::new(), &ast, name, (p1, p2))
+        engine.call_fn(&mut Scope::new(), &ast, name, (p1, p2))
     };
 
     // Test closure