+++ RELEASES.md
@@ -13,11 +13,13 @@ Breaking changes
 * `Engine::load_package` is renamed `Engine::register_global_module` and now must explicitly pass a shared [`Module`].
 * `Engine::register_module` is renamed `Engine::register_static_module` and now must explicitly pass a shared [`Module`].
 * `Package::get` is renamed `Package::as_shared_module`.
+* `Engine::set_module_resolver` now takes a straight module resolver instead of an `Option`. To disable module resolving, use the new `DummyModuleResolver`.
 
 Enhancements
 ------------
 
 * `Scope` is now `Clone + Hash`.
+* `Engine::register_static_module` now supports sub-module paths (e.g. `foo::bar::baz`).
 
 
 Version 0.19.8